sceripts/work.sh: update ws name generation
when opening a new work workspace, check if
there are existing workspaces of the desired name.
The list of existing workspaces is filtered to contain
only workspaces which names start with the desired name,
the number of found workspaces id incremented and
appended to the new workspace name.

e.g. one ws named "bla" exists, new ws "bla" will be called "bla2"

Signed-off-by: einKnie <einKnie@gmx.at>